Refactor map creation logic into a separate module.
Moved `creer_carte` and related functions to `utils/creer_carte.py` for better modularity and maintainability. Updated API key fetching to securely use `st.secrets`.
